+++ sentry/src/main/java/io/sentry/SentryEnvelopeItemHeader.java
@@ -24,7 +24,8 @@ public final class SentryEnvelopeItemHeader {
     if (getLength != null) {
       try {
         return getLength.call();
-      } catch (Exception ignored) {
+      } catch (Throwable ignored) {
+        // TODO: Take ILogger via ctor or hook into internal static logger
         return -1;
       }
     }